Move comparisons out of helpers
