+++ src/matchbox/client/helpers/__init__.py
@@ -1,13 +1,9 @@
 """Core functionalities of the Matchbox client."""
 
-from matchbox.client.helpers.comparison import comparison
 from matchbox.client.helpers.delete import delete_resolution
 from matchbox.client.helpers.selector import clean
 
 __all__ = (
-    # Sources
-    # Comparisons
-    "comparison",
     # Deletion
     "delete_resolution",
     # Selectors
