+++ src/matchbox/client/models/__init__.py
@@ -1,5 +1,10 @@
 """Deduplication and linking methodologies."""
 
+from matchbox.client.models.comparison import comparison
 from matchbox.client.models.models import Model, add_model_class
 
-__all__ = ("Model", "add_model_class")
+__all__ = (
+    "comparison",
+    "Model",
+    "add_model_class",
+)
