+++ src/matchbox/client/models/comparison.py
