+++ src/matchbox/client/models/linkers/deterministic.py
@@ -5,7 +5,7 @@ from collections.abc import Iterable
 import polars as pl
 from pydantic import Field, field_validator
 
-from matchbox.client.helpers import comparison
+from matchbox.client.models import comparison
 from matchbox.client.models.linkers.base import Linker, LinkerSettings
