+++ src/matchbox/client/models/linkers/weighteddeterministic.py
@@ -4,7 +4,7 @@ import duckdb
 import polars as pl
 from pydantic import BaseModel, Field, field_validator
 
-from matchbox.client.helpers import comparison
+from matchbox.client.models import comparison
 from matchbox.client.models.linkers.base import Linker, LinkerSettings
